UPDATE: Added Reel Functionality to the Harpoon Gun
Added a quick time event that allows the player to reel in enemies. Also added the reel in function and button map to the FPSActions map. Reel still needs to pull enemies towards player as well as render a line between the harpoon and player. Also needs fish-o-pedia functionality when that is added.