UPDATE: Harpoon System Overhaul & Grapple Shot
Massive overhaul to the harpoon system. The harpoon now has a state machine that can grab enemies and grapple points. When the player hits a grapplepoint with a harpoon, they'll now fly in the direction of the grapplepoint and unhook from it when they get within 1 unit of the grapplepoint.