UPDATE (FPS): Player health now has a "TakeKnockback" function, Urchin can now knock the player around when attacking.
Added a function to the FPPlayerHealth script that allows the player to take knockback from certain enemy attacks. The FPMove script now tracks the health script's "takingKnockback" bool, and doesn't allow movement inputs while the knockback is active. In addition, the urchin enemy now knocks the player backwards when landing an attack.