UPDATE (FPS and Tag manager): Added more placeholder tags, began making the logic for the piston.
Added 5 more dummy tags for jack and austin to the tag manager, made a script called HarpoonSwitch for detecting when an environmental hazard is hit by a harpoon projectile, and a FPPiston script for moving the piston. Also added logic to the collision detection of HarpoonController to trigger the "activate" and "deactivate" functions of different HarpoonSwitch objects.