UPDATE (FPS): Added the new Piranha enemy and various environmental detectors
The piranha enemy chases the player and does a fast strike. Also added a block that acts as an instant death zone when touched by the player or an enemy. Also added a block that turns off an enemy's navagent when they collide with it, so they don't respawn when they're flying over a bottomless pit. Changed the name of "EnemyHealthandQTE" to just "EnemyHealth", and adjusted all damage and health across the player and enemies to allow for much more higher damage both from and to the player.